refactor(dashboard): use Float32 type for AIOps float fields to avoid refresh drift
Adopt the established schema.Float32Attribute + types.Float32 pattern (used by
the time_slider panel) for random_sampler_probability and max_series_to_plot,
so float32 API fields round-trip without plan drift. Add panelkit.PreserveFloat32
helper. Also drop a redundant dead TimeRange assignment in the log rate import
branch (review feedback).

Refs #4005

+++ internal/kibana/dashboard/models/aiops.go
@@ -37,7 +37,7 @@ type AiopsPatternAnalysisConfigModel struct {
 	FieldName                types.String    `tfsdk:"field_name"`
 	MinimumTimeRange         types.String    `tfsdk:"minimum_time_range"`
 	RandomSamplerMode        types.String    `tfsdk:"random_sampler_mode"`
-	RandomSamplerProbability types.Float64   `tfsdk:"random_sampler_probability"`
+	RandomSamplerProbability types.Float32   `tfsdk:"random_sampler_probability"`
 	Title                    types.String    `tfsdk:"title"`
 	Description              types.String    `tfsdk:"description"`
 	HideTitle                types.Bool      `tfsdk:"hide_title"`
@@ -52,7 +52,7 @@ type AiopsChangePointChartConfigModel struct {
 	AggregationFunction types.String    `tfsdk:"aggregation_function"`
 	SplitField          types.String    `tfsdk:"split_field"`
 	Partitions          types.Set       `tfsdk:"partitions"`
-	MaxSeriesToPlot     types.Float64   `tfsdk:"max_series_to_plot"`
+	MaxSeriesToPlot     types.Float32   `tfsdk:"max_series_to_plot"`
 	ViewType            types.String    `tfsdk:"view_type"`
 	Title               types.String    `tfsdk:"title"`
 	Description         types.String    `tfsdk:"description"`
